# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API ",
        "description": "Listado de los productos extras",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/Playera": {
            "get": {
                "tags": [
                    "Playeras"
                ],
                "summary": "Consultar todas las playera",
                "description": "Retorna todas las playera disponibles",
                "operationId": "f8d863c39c70b5fe10234d290a0d0004",
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "Playera roja"
                                                    },
                                                    "tamaño": {
                                                        "type": "string",
                                                        "example": "L"
                                                    },
                                                    "color": {
                                                        "type": "string",
                                                        "example": "rojo"
                                                    },
                                                    "precio": {
                                                        "type": "number",
                                                        "example": "15"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Playeras"
                ],
                "summary": "Crear una nueva Playera",
                "description": "Crea una nueva Playera en el sistema.",
                "operationId": "556e817f2f38e89de0269cfa98ee3be7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Nombre",
                                    "Color",
                                    "Tamaño",
                                    "Precio",
                                    "Stock"
                                ],
                                "properties": {
                                    "Nombre": {
                                        "type": "string",
                                        "example": "Playera Azul"
                                    },
                                    "Color": {
                                        "type": "string",
                                        "example": "Azul"
                                    },
                                    "Tamaño": {
                                        "type": "string",
                                        "example": "M"
                                    },
                                    "Precio": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 15.99
                                    },
                                    "Stock": {
                                        "type": "integer",
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Playera creada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "Nombre": {
                                            "type": "string",
                                            "example": "Playera Azul"
                                        },
                                        "Color": {
                                            "type": "string",
                                            "example": "Azul"
                                        },
                                        "Tamaño": {
                                            "type": "string",
                                            "example": "M"
                                        },
                                        "Precio": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 15.99
                                        },
                                        "Stock": {
                                            "type": "integer",
                                            "example": 100
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solicitud incorrecta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/Playera/{id}": {
            "put": {
                "tags": [
                    "Playeras"
                ],
                "summary": "Actualizar una Playera existente",
                "description": "Actualiza los detalles de una Playera con el ID proporcionado.",
                "operationId": "b862239d23a5737f930cbf0c3c3da6b1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Nombre",
                                    "Color",
                                    "Tamaño",
                                    "Precio",
                                    "Stock"
                                ],
                                "properties": {
                                    "Nombre": {
                                        "type": "string",
                                        "example": "Playera Roja"
                                    },
                                    "Color": {
                                        "type": "string",
                                        "example": "Rojo"
                                    },
                                    "Tamaño": {
                                        "type": "string",
                                        "example": "L"
                                    },
                                    "Precio": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 20.99
                                    },
                                    "Stock": {
                                        "type": "integer",
                                        "example": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Playera actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "Nombre": {
                                            "type": "string",
                                            "example": "Playera Roja"
                                        },
                                        "Color": {
                                            "type": "string",
                                            "example": "Rojo"
                                        },
                                        "Tamaño": {
                                            "type": "string",
                                            "example": "L"
                                        },
                                        "Precio": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 20.99
                                        },
                                        "Stock": {
                                            "type": "integer",
                                            "example": 50
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontró la Playera",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Playera] #id"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Playeras"
                ],
                "summary": "Eliminar una Playera",
                "description": "Elimina una Playera del sistema usando el ID proporcionado.",
                "operationId": "164aaf80c86812bb8ce639a524b39885",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Playera eliminada correctamente"
                    },
                    "404": {
                        "description": "No se encontró la Playera",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Playera] #id"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Sanctum bearer token",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "bearer_token": []
        }
    ]
}